*: update logrus and sys/unix
As discussed in https://github.com/sirupsen/logrus/issues/570 and
various other issues, having multiple differently-cased versions of a
library around is a recipe for disaster.

Everyone else has converged on the new name for logrus, so we should
join in for the sake of any other dependencies we pull using it too.

This also bumps the unix syscall package since it was mildly stale.